refactor(agent-manager): organize Codex provider code
