+++ packages/agent-manager/src/index.ts
@@ -76,16 +76,16 @@ export type {
     ClaudePrintAgentServiceOptions,
     ClaudePrintSendResult,
 } from './providers/claude/durable/ClaudePrintAgentService.js';
-export { CodexCliProbe } from './durable/CodexCliProbe.js';
-export type { CodexCliProbeOptions } from './durable/CodexCliProbe.js';
-export { CodexPrintRunner } from './durable/CodexPrintRunner.js';
+export { CodexCliProbe } from './providers/codex/durable/CodexCliProbe.js';
+export type { CodexCliProbeOptions } from './providers/codex/durable/CodexCliProbe.js';
+export { CodexPrintRunner } from './providers/codex/durable/CodexPrintRunner.js';
 export type {
     CodexPrintRunnerOptions,
     CodexPrintRunRequest,
     CodexPrintRunResult,
-} from './durable/CodexPrintRunner.js';
-export { CodexPrintAgentService } from './durable/CodexPrintAgentService.js';
+} from './providers/codex/durable/CodexPrintRunner.js';
+export { CodexPrintAgentService } from './providers/codex/durable/CodexPrintAgentService.js';
 export type {
     CodexPrintAgentServiceOptions,
     CodexPrintSendResult,
-} from './durable/CodexPrintAgentService.js';
+} from './providers/codex/durable/CodexPrintAgentService.js';
