+++ packages/agent-manager/src/providers/codex/durable/CodexCliProbe.ts
@@ -1,6 +1,6 @@
 import { execFile } from 'child_process';
 import { promisify } from 'util';
-import { CodexPrintError } from './DurableAgent.js';
+import { CodexPrintError } from '../../../durable/DurableAgent.js';
 
 type ExecResult = { stdout: string; stderr: string };
 type Exec = (file: string, args: string[]) => Promise<ExecResult>;
